server/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
pymongo>=4.8.0
python-dotenv>=1.0.0
resend>=2.0.0